pyproject.toml
[project]
name = "openstack-k8s-modernization-lab"
version = "0.1.0"
description = "Local-first OpenStack + Kubernetes modernization engagement simulator (assessment, design, automation, KPIs)."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Daniel Alvarez" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.1",
  "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12.20240808",
]

[project.scripts]
okml = "okml.cli:app"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
pretty = true

[[tool.mypy.overrides]]
module = ["yaml.*", "matplotlib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=okml --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]

